requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-dotenv
requests
apscheduler
openai
twilio
email-validator
slowapi
bcrypt
PyJWT